Add lots of NrunRunners, add WOpen, add TARank.
Added lots of NrunRunnerTypes | Part1
(There are still several of them missing, maybe I'll continue another day adding them. Please, if you see an NrunRunnerType named incorrectly fix it or communicate it to someone so it can be fixed.)

Added WOpenPacket
(Pretty amazing packet that the server sends to the client for various reasons, uses WindowTypes for the different types of windows that can be pooped out).

Added WindowTypes
(Extremely related with UpgradePacketTypes, but still more or less independent. If there is some of them missing please add it or if you can ascertain the UpgradeSpProtected do it).

Added TARankPacket
(The structure of this packet is pretty large, so I tried to do my best doing the packet. You can see that in the PacketIndex16 there is a string. It takes the datetime formated as string (yyMMddhhmm) of when the last month winners were fixed as winners)

Added TARankSubPacket
(Used by the TARankPacket to specify a character. If you are going to use this SubPacket to make a packet near perfection I would watchout with the PacketIndex3, as when it is null it will output "-" and that should be "-1" because Official NosTale wants.)

"Added" some CancelPacketTypes.
(Just btw)